separate common resource and test functions
